Laser Pointer First Implementation
This is the first implementation of the laser pointer. You can hold right click anywhere on the webpage to switch your cursor to a laser pointer. If you move your mouse, a laser path is traced behind it. If you draw a circle or shake your mouse horizontally (pre-programmed gestures), you'll see a console log message detailing which shape you drew.